Damageable.isGoalTarget 추가 - 훈련용 봇을 플레이어로 취급하며, 점수를 지급할 수 있음

+++ src/main/java/com/dace/dmgr/combat/entity/temporary/Barrier.java
@@ -28,7 +28,7 @@ public abstract class Barrier extends SummonEntity<ArmorStand> implements Damage
     @NonNull
     protected final StatusEffectModule statusEffectModule;
     /** 죽었을 때 공격자에게 주는 점수 */
-    private final double score;
+    private final int score;
 
     /**
      * 방벽 인스턴스를 생성한다.
